+++ src/Core/src/Credentials/ChainProvider.php
@@ -5,6 +5,13 @@ declare(strict_types=1);
 namespace AsyncAws\Core\Credentials;
 
 use AsyncAws\Core\Configuration;
+use AsyncAws\Core\Exception\InvalidArgument;
+use Psr\Log\LoggerInterface;
+use Psr\Log\NullLogger;
+use Symfony\Component\Cache\Adapter\FilesystemAdapter;
+use Symfony\Component\HttpClient\HttpClient;
+use Symfony\Contracts\Cache\CacheInterface;
+use Symfony\Contracts\HttpClient\HttpClientInterface;
 use Symfony\Contracts\Service\ResetInterface;
 
 /**
@@ -59,5 +66,26 @@ final class ChainProvider implements CredentialProvider, ResetInterface
     public function reset()
     {
         $this->lastSuccessfulProvider = [];
+    }
+
+    public static function createDefaultChain(?HttpClientInterface $httpClient = null, ?LoggerInterface $logger = null, $cache = null): CredentialProvider
+    {
+        $httpClient = $httpClient ?? HttpClient::create();
+        $logger = $logger ?? new NullLogger();
+        if (null === $cache) {
+            if (\interface_exists(CacheInterface::class)) {
+                $cache = new FilesystemAdapter('async-aws');
+            }
+        } elseif (!$cache instanceof CacheInterface) {
+            throw new InvalidArgument(sprintf('Expected cache to be a "%s". Got "%s".', CacheInterface::class, \get_class($cache)));
+        }
+
+        return new CacheProvider(new ChainProvider([
+            new ConfigurationProvider(),
+            new WebIdentityProvider($logger),
+            new IniFileProvider($logger),
+            null === $cache ? new ContainerProvider($httpClient, $logger) : new PersistingCacheProvider(new ContainerProvider($httpClient, $logger), $cache),
+            null === $cache ? new InstanceProvider($httpClient, $logger) : new PersistingCacheProvider(new InstanceProvider($httpClient, $logger), $cache),
+        ]));
     }
 }